requirements.txt
# requirements.txt ── все зависимости для запуска приложения

Django==5.0.4
djangorestframework==3.15.1
djoser==2.2.2
drf-extra-fields==3.6.1
django-filter==23.2

Pillow==10.2.0
psycopg[binary]==3.1.18
python-dotenv==1.0.1
reportlab==4.0.0
gunicorn==20.1.0
